Refactor GQL work facets into an interface
