+++ app/graphql/types/interfaces/base_interface.rb
@@ -1,0 +1,3 @@
+module BaseInterface
+  include GraphQL::Schema::Interface
+end
