+++ config/application.rb
@@ -78,9 +78,12 @@ module Lupo
                      eager_load: true
     config.paths.add Rails.root.join("app", "graphql", "resolvers").to_s,
                      eager_load: true
+    #config.paths.add Rails.root.join("app", "graphql", "types", "interfaces").to_s,
+                     #eager_load: true
 
     # Allow middleware to be loaded. (compressed_requests)
     config.autoload_paths += %W(#{config.root}/lib #{config.root}/lib/middleware)
+    config.autoload_paths += %W(#{config.root}/app/graphql/types )
     config.eager_load_paths += %W(#{config.root}/lib #{config.root}/lib/middleware)
 
     # Settings in config/environments/* take precedence over those specified here.
